1. Add exceptions to LICENSE; 2. Change for loops to meet the С++ standard; 3. Fix memory leak in mtrc-object.cpp.

+++ nodes/all_nodes.h
@@ -26,7 +26,7 @@ private:
 class subprogram_unit : public code_block
 {
 public:
-	int getOffset(const unsigned char* current) const { return ((int)this->address) - ((int)current); }
+	int getOffset(const unsigned char* current) const { return static_cast<int>(this->address - current); }
 protected:
 	unsigned char* address;
 };
